+++ packages/aesthetic-adapter-aphrodite/tests/UnifiedAdapter.test.js
@@ -15,6 +15,7 @@ import {
   SYNTAX_IMPORT,
   SYNTAX_KEYFRAMES,
   SYNTAX_MEDIA_QUERY,
+  SYNTAX_MULTI_SELECTOR,
   SYNTAX_NAMESPACE,
   SYNTAX_PAGE,
   SYNTAX_PROPERTIES,
@@ -124,6 +125,21 @@ describe('aesthetic-adapter-aphrodite/UnifiedAdapter', () => {
     expect(instance.syntax.convert(SYNTAX_PSEUDO)).toEqual(SYNTAX_PSEUDO);
 
     expect(instance.transform(instance.create(SYNTAX_PSEUDO).pseudo)).toBe('pseudo_q2zd6k');
+
+    expect(renderAphroditeStyles(instance)).toMatchSnapshot();
+  });
+
+  it('handles multiple selectors (comma separated)', () => {
+    expect(instance.syntax.convert(SYNTAX_MULTI_SELECTOR)).toEqual({
+      multi: {
+        cursor: 'pointer',
+        ':disabled': { cursor: 'default' },
+        '[disabled]': { cursor: 'default' },
+        '> span': { cursor: 'default' },
+      },
+    });
+
+    expect(instance.transform(instance.create(SYNTAX_MULTI_SELECTOR).multi)).toBe('multi_1xnvdcd');
 
     expect(renderAphroditeStyles(instance)).toMatchSnapshot();
   });