+++ packages/aesthetic-adapter-typestyle/CHANGELOG.md
@@ -1,9 +1,10 @@
-# 1.1.0
+# 1.1.0 - 01/13/18
 #### 🚀 New
 * Can now pass dynamic style objects to `transform` (will be akin to inline).
 * Unified Syntax
   * Attribute selectors are now supported when starting with `[`.
   * Direct descendant selectors are now supported with `>`.
+  * Multiple selectors are now supported by separating each selector with a comma.
 
 #### 🐞 Fixed
 * Resolved some issues with pseudo classes.